+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.0
+* refactor code, remove `API` and `Auth` class, replace with `MiCloud` class.
+* support more features: get device list, get device data, get device props, set device props, call device action.
+
 ## 0.2.0
 * Support more countries.